[changed] Split <Routes> component from route spec
Fixes #112.

Test Plan: Loaded each of the examples and clicked around. npm test too.
