+++ examples/data-flow/app.js
@@ -2,6 +2,7 @@
 var React = require('react');
 var Router = require('../../modules/main');
 var Route = Router.Route;
+var Routes = Router.Routes;
 var Link = Router.Link;
 
 var App = React.createClass({
@@ -64,9 +65,11 @@ var Taco = React.createClass({
 });
 
 var routes = (
-  <Route handler={App}>
-    <Route name="taco" path="taco/:name" handler={Taco}/>
-  </Route>
+  <Routes>
+    <Route handler={App}>
+      <Route name="taco" path="taco/:name" handler={Taco}/>
+    </Route>
+  </Routes>
 );
 
 React.renderComponent(routes, document.body);
