+++ examples/query-params/app.js
@@ -2,6 +2,7 @@
 var React = require('react');
 var Router = require('../../modules/main');
 var Route = Router.Route;
+var Routes = Router.Routes;
 var Link = Router.Link;
 
 var App = React.createClass({
@@ -32,9 +33,11 @@ var User = React.createClass({
 });
 
 var routes = (
-  <Route handler={App}>
-    <Route name="user" path="user/:userId" handler={User}/>
-  </Route>
+  <Routes>
+    <Route handler={App}>
+      <Route name="user" path="user/:userId" handler={User}/>
+    </Route>
+  </Routes>
 );
 
 React.renderComponent(routes, document.body);
