+++ examples/shared-root/index.html
@@ -1,4 +1,5 @@
 <!doctype html public "superawesome">
 <title>Shared Root Example</title>
+<link rel="stylesheet" href="../app.css"/>
 <body>
 <script src="../build/shared-root.js"></script>
