+++ examples/simple-master-detail/app.js
@@ -2,6 +2,7 @@
 var React = require('react');
 var Router = require('../../modules/main');
 var Route = Router.Route;
+var Routes = Router.Routes;
 var Link = Router.Link;
 
 var App = React.createClass({
@@ -50,9 +51,11 @@ var State = React.createClass({
 });
 
 var routes = (
-  <Route handler={App}>
-    <Route name="state" path="state/:abbr" handler={State}/>
-  </Route>
+  <Routes>
+    <Route handler={App}>
+      <Route name="state" path="state/:abbr" handler={State}/>
+    </Route>
+  </Routes>
 );
 
 React.renderComponent(routes, document.body);
