+++ examples/transitions/app.js
@@ -2,6 +2,7 @@
 var React = require('react');
 var Router = require('../../modules/main');
 var Route = Router.Route;
+var Routes = Router.Routes;
 var Link = Router.Link;
 
 var App = React.createClass({
@@ -55,10 +56,12 @@ var Form = React.createClass({
 });
 
 var routes = (
-  <Route handler={App}>
-    <Route name="dashboard" handler={Dashboard}/>
-    <Route name="form" handler={Form}/>
-  </Route>
+  <Routes>
+    <Route handler={App}>
+      <Route name="dashboard" handler={Dashboard}/>
+      <Route name="form" handler={Form}/>
+    </Route>
+  </Routes>
 );
 
 React.renderComponent(routes, document.body);
