+++ modules/stores/RouteStore.js
@@ -1,5 +1,6 @@
 var React = require('react');
 var invariant = require('react/lib/invariant');
+var warning = require('react/lib/warning');
 var Path = require('../helpers/Path');
 
 var _namedRoutes = {};
